+++ base/deprecated.jl
@@ -3423,6 +3423,9 @@ workspace() = error("workspace() is discontinued, check out Revise.jl for an alt
 # PR #25113
 @deprecate_binding CartesianRange CartesianIndices
 
+@deprecate tryparse(T::Type, str::AbstractString) parse(Union{T, Void}, str)
+@deprecate tryparse(T::Type, str::AbstractString, base::Integer) parse(Union{T, Void}, str, base)
+
 # END 0.7 deprecations
 
 # BEGIN 1.0 deprecations