+++ base/exports.jl
@@ -315,7 +315,6 @@ export
     fldmod1,
     flipsign,
     float,
-    tryparse,
     floor,
     fma,
     frexp,