+++ NEWS.md
@@ -840,6 +840,9 @@ Deprecated or removed
 
   * `sub2ind` and `ind2sub` are deprecated in favor of using `CartesianIndices` and `LinearIndices` ([#24715]).
 
+  * `tryparse(T, s[, base])` has been deprecated in favor of `parse(Union{T, Void}, s[, base])`
+    ([#25130]).
+
 Command-line option changes
 ---------------------------
 
@@ -1065,3 +1068,4 @@ Command-line option changes
 [#25056]: https://github.com/JuliaLang/julia/issues/25056
 [#25057]: https://github.com/JuliaLang/julia/issues/25057
 [#25088]: https://github.com/JuliaLang/julia/issues/25088
+[#25130]: https://github.com/JuliaLang/julia/issues/25130